Also link labels to https://en.wiktionary.org/ when the user answered incorrectly. Closes #169.
