+++ src/toisto/ui/text.py
@@ -1,16 +1,17 @@
 """Output for the user."""
 
 from datetime import datetime, timedelta
-from string import punctuation
 from typing import NoReturn
 import sys
 
 from rich.console import Console
 from rich.theme import Theme
 
-from .diff import colored_diff
 from ..metadata import NAME, VERSION
 from ..model import Label, Quiz
+
+from .diff import colored_diff
+from .dictionary import linkify_and_enumerate
 
 
 theme = Theme(dict(secondary="grey69", quiz="medium_purple1", inserted="bright_green", deleted="bright_red"))
@@ -62,28 +63,6 @@ def format_duration(duration: timedelta) -> str:
 def format_datetime(date_time: datetime) -> str:
     """Return a human readable version of the datetime"""
     return date_time.isoformat(sep=" ", timespec="minutes")
-
-
-def split_punctuation(text: str) -> tuple[str, str, str]:
-    """Return a tuple of prefixed punctuation, the text without punctuation, and postfix punctuation."""
-    stripped_text = text.strip(punctuation)
-    stripped_text_start = text.find(stripped_text)
-    stripped_text_end = stripped_text_start + len(stripped_text)
-    return text[:stripped_text_start], text[stripped_text_start:stripped_text_end], text[stripped_text_end:]
-
-
-def linkify(text: str) -> str:
-    """Return a version of the text where each word is turned into a link to a dictionary."""
-    linkified_words = []
-    for word in text.split():
-        prefix, word, postfix = split_punctuation(word)
-        linkified_words.append(f"{prefix}[link=https://en.wiktionary.org/wiki/{word.lower()}]{word}[/link]{postfix}")
-    return " ".join(linkified_words)
-
-
-def linkify_and_enumerate(*texts: str, sep: str = ", ") -> str:
-    """Return a linkified and enumerated version of the texts."""
-    return sep.join(f'"{linkify(text)}"' for text in texts)
 
 
 def feedback_correct(guess: Label, quiz: Quiz) -> str:
