+++ tests/toisto/command/test_practice.py
@@ -5,7 +5,8 @@ from unittest.mock import call, patch, Mock, MagicMock
 from toisto.command import practice
 from toisto.model import Progress, Topic, Topics
 from toisto.model.model_types import ConceptId
-from toisto.ui.text import DONE, TRY_AGAIN, linkify
+from toisto.ui.dictionary import linkify
+from toisto.ui.text import DONE, TRY_AGAIN
 
 from ..base import ToistoTestCase
